+++ src/main/java/boguszGroup/Blog/utils/Ids.java
@@ -4,12 +4,12 @@ import java.util.Set;
 
 public class Ids {
 
- static public int generateNewId(Set<Integer> keysSoFar){
+ static public long generateNewId(Set<Long> keysSoFar){
     if(keysSoFar.isEmpty()){
       return 0;
     }
     else{
-      Integer integer = keysSoFar.stream().max((o1, o2) -> o1.compareTo(o2)).get();
+      long integer = keysSoFar.stream().max((o1, o2) -> o1.compareTo(o2)).get();
       return integer+1;
     }
   }
